// src/interfaces.ts
export interface ICard {
  category: string;
  type: string;
  name: string;
  desc: string;
  rarity?: string;
  cost?: string;
}

// export enum cardBase {
//   'ironclad',
//   'colorless',
//   'status',
//   'curse',
// }

// export enum cardRarity {
//   'common',
//   'uncommon',
//   'rare',
//   'special',
// }

// export enum cardType {
//   'attack',
//   'power',
//   'skill',
// }